[generators] Rework namespaces support
It seems the order of the elements returned from `Declaration.
GatherNamespaces` can vary, so removing the `First` element can
be wrong (leading to an incorrect.missing namespace), crashing
the app (since the class won't be found).

Replace this with a loop that exclude the `TranslationUnit` (a
subclass of `Namespace`) and use a `StringBuilder` to minimize
allocations.