[objc] When init returns null, self-destruction is required.
When init returns null, we have to self-destruct, since the caller won't be
able to destroy us.

Since this fixes a memory leak, objects that failed to initialize are now
deallocated properly, and we need to check in the dealloc method if the
managed object really exists before destroying it.