[objc] Support ValueType with no default constructor
This happens if there's no initialization. Previously the lack of a
public `.ctor()` meant we did not generate the code (nothing to call).

This was not correct (since it's valid in .net) so we detect the
condition and generate a simpler `init` method in such case.

Unit tests added (for the existing `Point` type that suffered, unnoticed,
from this)