[objc] Make sure we can build for iOS release using mtouch. Partial fix for #237
Embedding with Xamarin.iOS is different than doing it directly with mono.
This requires some methods to be defined, like the ones only used for
release (when `TOKENLOOKUP` is defined).

The integration with mtouch is also incomplete (for non-debug build) as
some steps (MdbSym) and SymbolStrip (only used in release by default)
causes FileNotFoundException (some state is not set correctly).

This PR is a fix for the former (headers), a workaround for the later
(mtouch invocation) and a Makefile change to make sure we always build
release/iOS on bots (because untested stuff never works, for long)

ref: https://github.com/mono/Embeddinator-4000/issues/237