[objc] Add an assembly load hook, and use it to find assemblies in bundles.
Remove the assembly search hook that returns the path of an assembly, and
instead have an assembly load hook, which returns an already loaded
MonoAssembly *.

This makes writing the hook a bit simpler, because string-returning hooks need
specific memory management (a hook that returns a "const char *" won't be able
to use dynamically allocated strings easily, and returning a "char *" runs
into questions about how's to free that string). It's much simpler to just
return a MonoAssembly*, which mono must free when unloaded.