[objc] Add support for System.TimeSpan into generated projects (#327)
This replace the older (not linker friendly) PR #251.

It brings, if required, `System.TimeSpan` inside your projects since
there is direct, good alternative, i.e. `NSTimeInterval` is just a
typedef on `double`.

* [mac] Link SDK is only supported for the modern profile

* [mac] Provide `--nolink` to mmp when the linker is not supported

The earlier problem (and test failures) is that, by default `mmp` (like
`mtouch`) does a **link all**.

It's easy to forget since we don’t normally use the command line (but
the UI of an IDE) to call `mmp`.