[objc] Add `--nowarn:` and `--warnaserror:` options to the driver
Code was already present (copied from mtouch/mmp with ErrorHelper.cs) so
it's an easy and useful thing to expose.

Use `--nowarn` on UnitsNet sample to reduce the warnings that we cannot
fix (since we're using a binary downloaded from nuget)

Add unit tests for `--nowarn`. The one for `--warnaserror` did not work,
at least inside VSfM, due to bug #55801, resulting in:

> mono_os_mutex_lock: pthread_mutex_lock failed with "Invalid argument" (22)