[c] [java] fix for properties of type UnsupportedType
In testing various libraries, I’ve discovered a subclass of
`System.EventArgs` always breaks the C and Java generators. The
property `EventArgs.Empty` ends up as an `UnsupportedType`, which in
turn prints an empty string for its type.

Skipping these properties is a solution for now, it is at least better
than hitting a C or Java compiler error.

Changes:
- Added integration test that reproduces the issue
- C generator skips `UnsupportedType` for headers & sources
- Java generator skips `UnsupportedType` (only sources)