[build] remove anything related to AppVeyor (#666)
As part of consolidating our CI setup, we now have Windows builds running on Jenkins. This means we no longer need to use AppVeyor at all, which is a good thing. AppVeyor is not a CI system well known by other folks working on the Xamarin or Mono teams, where Jenkins is mostly used.

Other changes:
- Updated `.gitignore` for files appearing on Windows: `UpgradeLog.htm` from VS upgrading projects and `*.obj` from running the C tests